pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "onnx_clip"
version = "3.0.0"
description = "CLIP with ONNX Runtime and without PyTorch dependencies."
readme = "README.md"
authors = ["Lakera AI <dev@lakera.ai>"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/lakeraai/onnx_clip"

[tool.poetry.dependencies]
python = "^3.6.1"
numpy = "^1.18.0"
opencv-python-headless = "^4.0.1"
onnxruntime = ">=1.4.0"
pillow = "^8.4.0"
ftfy = "^6.0.3"
regex = "*"
boto3 = "^1.23.10"
importlib_metadata = ">=4.8"  # Supported by Python 3.6